refactor(node): share require call tracking in nodeutil
